+++ index.html
@@ -28,7 +28,7 @@
     </div>
   </div>
   <script src="https://cdn.jsdelivr.net/npm/@woowacourse/mission-utils@1.0.1/dist/mission-utils.min.js"></script>
-  <script type="module" src="src/index.js"></script>
+  <script type="module" src="./src/index.js"></script>
 </body>
 
 </html>
